requirements.txt
faiss-cpu
pandas
numpy
sentence-transformers
